feat: create helper to generate location description andd worker label
