+++ ui/src/components/entities-table/utils.ts
@@ -5,9 +5,7 @@ import type {
   EntityListRequest,
   EntityListResponse,
   FiniteStateMachine,
-  Operator,
   OperatorFilter,
-  Plan,
   QueryFilter,
   Resource,
   ResourceGroup,
@@ -165,22 +163,6 @@ export function parseOptionalNumber(value: string): number | null {
   }
   const parsed = Number(value);
   return Number.isFinite(parsed) ? parsed : null;
-}
-
-/** Builds a "Plan / Worker" subtitle so operators sharing the same name can be told apart. */
-export function operatorLocationDescription(
-  operator: Operator,
-  plans: Record<string, Plan>,
-  workers: Record<string, Worker>
-): string | undefined {
-  const plan = operator.plan_id ? plans[operator.plan_id] : undefined;
-  if (!plan) {
-    return undefined;
-  }
-  const planLabel = plan.instance_name ?? plan.id;
-  const worker = plan.worker_id ? workers[plan.worker_id] : undefined;
-  const workerLabel = worker ? (worker.instance_name ?? worker.id) : null;
-  return workerLabel ? `Plan: ${planLabel} · Worker: ${workerLabel}` : `Plan: ${planLabel}`;
 }
 
 /** Builds a "Worker" subtitle so resources sharing the same name across workers can be told apart. */
